+++ CHANGELOG.md
@@ -21,6 +21,7 @@
 * Added Iterator class and Array#createIterator()
 * Added Deck class (ordered object)
 * Added unit tests for Iterator & Deck
+* Added String#dissect(open, close)
 
 ## 0.1.1 (2014-06-23)